third_party/WebKit/LayoutTests/fast/html/imports/import-owner.html

<!DOCTYPE html>
<html>
<script src="../../js/resources/js-test-pre.js"></script>
<head>
<link id="target1" rel="import" href="resources/hello.html">
<link id="target2" rel="import" href="resources/bye.html">
</head>
<body>
<script>
shouldBeNonNull("target1.import.ownerDocument");
shouldNotBe("target1.import.ownerDocument", "document");
shouldBe("target1.import.ownerDocument", "target2.import.ownerDocument");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
