third_party/WebKit/LayoutTests/fast/html/imports/import-owner-expected.txt

PASS target1.import.ownerDocument is non-null.
PASS target1.import.ownerDocument is not document
PASS target1.import.ownerDocument is target2.import.ownerDocument
PASS successfullyParsed is true

TEST COMPLETE
